requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
autopep8==2.0.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
cssselect2==0.7.0
Django==4.1.3
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
html5lib==1.1
idna==3.4
lxml==4.9.2
oscrypto==1.3.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.23.0
PyJWT==2.6.0
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.29.0
sib-api-v3-sdk==7.6.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
typing-extensions==4.5.0
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
xhtml2pdf==0.2.11
